% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_index.R
\name{age_index}
\alias{age_index}
\title{Age indices are assigned to the proxy data}
\usage{
age_index(
  load_proxy = load_foram,
  age_units = "kyr",
  step_type = "regular",
  step_int = 10,
  cust_steps
)
}
\arguments{
\item{load_proxy}{User inputs the list that is returned by the 'load_foram', 'load_phyto', 'load_paleosol', or 'load_plants'
functions. Defauts to 'load_foram'.}

\item{age_units}{Indicate the age units used in the input data. Either 'kyr' or 'Myr'. Defaults to 'kyr'.}

\item{step_type}{Indicate either 'regular' for regular interval time stepping with step interval specified,
i.e., 'step_int', OR 'every' for variable time steps at each age for which there is data, and 'step_int' is not required, 
OR 'both' to have time steps at each age for which there is data in addition to regular time step intervals - 'step_int' must
be specified for 'both' step_type argument) OR 'custom' to provide your own vector of time step ages ('cust_steps' vector 
of ages must be included as argument for 'custom' step_type). Defaults to 'regular'.}

\item{step_int}{If 'step_type' is 'regular' or 'both', indicated the 'step_int' or regular step interval to be used. 
Defaults to '10'.}

\item{cust_steps}{If 'custom' is specified for 'step_type' provide a vector of ages. There will be a time step for each 
age supplied. Defaults to null.}
}
\value{
Returns list 'age_index'. This list contains 1. 'prox_in_ai' data frame which has been age indexed by the function,
2. 'ages_prox' which is a vector of the ages of each model time step to be used, 3. 'dt' which is a vector of the delta time, or
amount of time between each time step, and 4. 'obs_type' carried over from input list.
}
\description{
User specifies time stepping to be used by model, either 'regular' for regular interval time stepping
with step interval specified, i.e., 'step_int', OR 'every' for variable time steps at each age for which there is data.
Each row of data is assigned an index which will be used in the time series model.
}
\examples{
age_index(load_proxy = load_foram, age_units = "kyr", step_type = "regular", step_int = 10)

}
