% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbchem.R
\name{carbchem}
\alias{carbchem}
\title{Marine carbonate chemistry calculator}
\usage{
carbchem(
  ccparm2 = "dic",
  ccout = "alk",
  inv_out,
  pH,
  ccparm2.vec,
  temp,
  sal,
  press,
  xca,
  xmg,
  xso4
)
}
\arguments{
\item{ccparm2}{User indicates which 2nd carbonate chemistry variable to use: 'dic', 'alk', 'pco2', 'co3', 'hco3', or 'omegac'.
Defaults to 'dic'.}

\item{ccout}{Specify which parameter to compute: 'dic', 'alk', 'pco2', 'co3', 'hco3', or 'omegac'. Defaults to 'pco2'.}

\item{inv_out}{Use inversion output returned object from 'run_inversion' function to compute carbonate chemistry. Using 'inv_out' 
alone for this calculation requires that all the variables that go into the carbonate chemistry calculation were included in the 
'save.parms' argument of the 'run_inversion' function. Otherwise, specify equal length vectors of values using individual variable 
arguments. Defaults to NULL.}

\item{pH}{Provide a vector of pH data in total scale if 'inv_out' is not used. Defaults to NULL.}

\item{ccparm2.vec}{Provide a vector of data of the type specified by 'ccparm2' if 'inv_out' is not used. Defaults to NULL.}

\item{temp}{Provide a vector of temperature data in degrees C if 'inv_out' is not used. Defaults to NULL.}

\item{sal}{Provide a vector of salinity data if 'inv_out' is not used. Defaults to NULL.}

\item{press}{Provide a vector of pressure data in bar if 'inv_out' is not used. Defaults to NULL.}

\item{xca}{Provide a vector of Ca ion concentration data in mmol/kg if 'inv_out' is not used. Defaults to NULL.}

\item{xmg}{Provide a vector of Mg ion concentration data in mmol/kg if 'inv_out' is not used. Defaults to NULL.}

\item{xso4}{Provide a vector of SO4 ion concentration data in mmol/kg if 'inv_out' is not used. Defaults to NULL.}
}
\value{
Returns vector of values for parameter of user interest specified by 'ccout' argument.
}
\description{
This function computes any 6 of the master marine carbonate chemistry variables, plus omega for calcite, from pH data
plus an additional carbonate chem variable data set. The user must suply equal length vectors for each influencing variable
in the calculations: 'pH', 2nd carb chem variable 'ccparm2.vec', 'temp', 'sal', 'press', 'xca', 'xmg', and 'xso4'.
Units for each variable as follows: pH = total scale; ccparm2 = mmol/kg (dic, alk, co3, hco3) OR ppmv (pco2) OR unitless (omegac);
temp = deg C, sal = unitless; press = bar; elemental concentrations are in mmol/kg.
}
\examples{
carbchem(ccparm2 = "dic", ccout = "alk", inv_out = inv_out)

}
