% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_foram.R
\name{clean_foram}
\alias{clean_foram}
\title{Formats foraminiferal observational data}
\usage{
clean_foram(age_index = age_index)
}
\arguments{
\item{age_index}{User inputs the list that is returned by the 'age_index' function. Defaults to 'age_index'.}
}
\value{
Returns list 'clean_obs'. This list contains objects associated with observational data, age
indexing and species indexing which will be used in the model inversion.
}
\description{
This function generates a list of observations, age indexes, and species indexes formatted for use in the model
inversion. It also species-indexes the d11B data which will allow for different 'vital effects' to be applied
to each sample data set if indicated in 'load_foram'.
}
\examples{
clean_foram(age_index = age_index)

}
