% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_persons.R
\docType{data}
\name{example_persons}
\alias{example_persons}
\title{Example Persons Data Set}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 13 columns.
}
\usage{
example_persons
}
\description{
All individuals, except the first row, are randomly generated from the
ethnorace distribution data sets.
}
\keyword{datasets}
