% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_ethnorace.R
\name{impute_ethnorace}
\alias{impute_ethnorace}
\title{Impute Ethnicity/Race}
\usage{
impute_ethnorace(input_data, bper_data = NULL, year, census_key = NULL)
}
\arguments{
\item{input_data}{The input data frame containing the individuals whose
ethnorace the user wants to impute. The following input variables are
supported: \code{last_name}, \code{first_name}, \code{age}, \code{sex}, \code{party}, \code{multi_unit},
\code{state}, \code{county}, \code{zip}, \code{place}, \code{tract}, \code{district}, \code{block}.}

\item{bper_data}{The input data list containing ethnorace conditional
probabilities. If left empty, will default to downloading directly from
Census API. Use the function \code{\link{load_bper_data}} to save this
input data list ahead of time.}

\item{year}{The year for which Census data will be loaded. The function will
retrieve the closest available Census data for the year.}

\item{census_key}{Personal Census API key. See
https://api.census.gov/data/key_signup.html}
}
\value{
The original data frame with the additional columns for ethnorace
probabilities and highest probability imputed category.
}
\description{
Calculates posterior probabilities for individual ethnorace categories using
the Naive Bayes algorithm. Also returns highest probability ethnorace as a
new character column in the data frame.
}
