% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_bper_data.R
\name{load_bper_data}
\alias{load_bper_data}
\title{Load data for \code{\link{impute_ethnorace}} function}
\usage{
load_bper_data(
  input_data,
  year,
  census_key = NULL,
  ll = list(last_name = 1, first_name = 1, state = 1, county = 1, zip = 1, place = 1,
    tract = 1, district = 1, block = 1),
  ...
)
}
\arguments{
\item{input_data}{The input data frame containing the individuals whose
ethnorace the user wants to impute. The following input variables are
supported: \code{last_name}, \code{first_name}, \code{age}, \code{sex}, \code{party}, \code{multi_unit},
\code{state}, \code{county}, \code{zip}, \code{place}, \code{tract}, \code{district}, \code{block}.}

\item{year}{The year for which Census data will be loaded. The function will
retrieve the closest available Census data for the year.}

\item{census_key}{Personal Census API key. See
https://api.census.gov/data/key_signup.html}
}
\value{
A list with data objects necessary for
\code{\link{impute_ethnorace}}.
}
\description{
This function retrieves the demographic data for use in
\code{\link{impute_ethnorace}}.
}
