% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_bperdata.R
\name{load_bperdata}
\alias{load_bperdata}
\title{Download Ethnorace Distribution Data}
\usage{
load_bperdata(
  destination = NA,
  download = FALSE,
  load_data = TRUE,
  save_files = TRUE
)
}
\arguments{
\item{destination}{A folder in the current working directory into which you
would like to download the data files. The function will create the folder
if no folder of that name currently exists. If left blank defaults to
downloading the files into the current working directory.}

\item{download}{Option to download data sets from
\url{https://github.com/bwilden/bperdata/tree/master/data}. Only needs to
be done once.}

\item{load_data}{Option to load data sets into the Global Environment? If you
want to download the files only, set 'load_data = FALSE'.}

\item{save_files}{Option to delete files in directory after loading into
Global Environment.}
}
\description{
This function downloads, and loads into the Global Environment, the set of
data files necessary for \code{\link{predict_race}}. Combined file size is 59
MB. For more information see \url{https://github.com/bwilden/bperdata}.
}
\section{Files}{
 \code{apartments.rda}, \code{birth_years.rda},
\code{blocks.rda}, \code{firstnames.rda}, \code{genders.rda},
\code{parties.rda}, \code{surnames.rda}, \code{zips.rda}
}

\examples{
# First time:

\dontrun{
load_bperdata(destination = "bperdata", download = TRUE)
}

# After files are downloaded:

\dontrun{
load_bperdata(destination = "bperdata", download = FALSE)
}

}
\seealso{
\code{\link{predict_race}}
}
