% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_plot.R
\name{post_plot}
\alias{post_plot}
\title{Plots time series parameter posteriors}
\usage{
post_plot(inv_out = inv_out, parm = "dic", type = "CI")
}
\arguments{
\item{inv_out}{User inputs the MCMC data object that is returned by the 'run_inversion' function. Defaults to 'inv_out'.}

\item{parm}{Specify which parameter to plot. This parameter must be included in 'save.parms' argument in 'run_inversion'
function. Defaults to 'pco2'.}

\item{type}{Indicate which plot style desired. Either 'CI' for 95 percent credible interval, or 'draws' for 500 draws of time
series posteriors. Defaults to 'CI'.}
}
\value{
Returns a plot called 'post_plot'.
}
\description{
This function generates a time series plot for a mcmc inversion output parameter. This parameter must be
included in 'save.parms' argument in 'run_inversion' function. Two plot style options are included.
}
\examples{
post_plot(inv_out = inv_out, parm = "pco2", type = "CI")

}
