% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_plot_ind.R
\name{post_plot_ind}
\alias{post_plot_ind}
\title{Plots parameter posterior for a single time step}
\usage{
post_plot_ind(inv_out = inv_out, parm = "pco2", tstep_age, show.median = FALSE)
}
\arguments{
\item{inv_out}{User inputs the MCMC data object that is returned by the 'run_inversion' function. Defaults to 'inv_out'.}

\item{parm}{Specify which parameter to plot. This parameter must be included in 'save.parms' argument in 'run_inversion'
function. Defaults to 'pco2'.}

\item{tstep_age}{Indicate which time step age is desired if the parameter is time-dependent. This age much match one of
the input ages, compiled in 'ages_prox' object. No default.}

\item{show.median}{Logical. Specify TRUE if you would like a line drawn through the median of the distribution. Defaults
to FALSE.}
}
\value{
Returns a plot 'post_plot_out'.
}
\description{
This function generates a time series plot for a mcmc inversion output parameter of interest. This parameter must be
included in 'save.parms' argument in 'run_inversion' function. Two plot style options are included.
}
\examples{
post_plot_ind(inv_out = inv_out, parm = "pco2", tstep_age, show.median = FALSE)

}
