% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ethnorace.R
\name{predict_ethnorace}
\alias{predict_ethnorace}
\title{Predict Ethnicity/Race}
\usage{
predict_ethnorace(df, dichotomize = FALSE)
}
\arguments{
\item{df}{A data frame containing some number of the following column names:
\code{last_name}, \code{first_name}, \code{female}, \code{apartment},
\code{party}, \code{birth_year}, \code{state}, \code{county}, \code{zip},
\code{block}. No individual variable, or set of variables, is necessary for
the algorithm to function, but including more information will yield better
predictions (excluding geolocation variables). For each individual, only
the most "granular" geolocation variable will be used.}

\item{dichotomize}{Indicates whether to create binary factor variables
indicating an individual's category.}
}
\value{
Returns the original data.frame with the additional columns for
ethnorace probabilities and predicted category.
}
\description{
Calculates posterior probabilities for individual ethnorace categories using
the Naive Bayes algorithm. Also returns highest predicted race as a new
string column in the data frame.
}
