% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_race.R
\name{predict_race}
\alias{predict_race}
\title{Predict Race}
\usage{
predict_race(df, dichotomize = FALSE)
}
\arguments{
\item{df}{A data frame containing the following column names:
\code{last_name}, \code{first_name}, \code{female} , \code{apartment},
\code{party}, \code{birth_year}, \code{zip}, \code{block}. These columns
need to all be included to run the function. If missing, please create a
column of NA's corresponding to the correct type.}

\item{dichotomize}{Indicates whether to create binary factor variables
indicating an individual's category. 'FALSE' by default.}
}
\value{
Returns the original data.frame with the additional columns for ethnorace
probabilities and predicted category.
}
\description{
Calculates posterior probabilities for individual ethnorace categories using
the Naive Bayes algorithm. Also returns highest predicted race as a new
string column in the data frame.
}
