% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors_foram.R
\name{priors_foram}
\alias{priors_foram}
\title{Loads prior distributions for foram model}
\usage{
priors_foram(
  parms_foram_adj,
  age_index = age_index,
  cc2ndparm.vt = "dic",
  cc2ndparm.pt = "ts",
  cc2ndparmTS
)
}
\arguments{
\item{parms_foram_adj}{Specify the R object or character string file path to configuration script used to define parameters 
and priors. Defaults to system data file if no object or path is provided.}

\item{age_index}{User inputs the list that is returned by the 'age_index'. Defaults to 'age_index'.}

\item{cc2ndparm.vt}{Second carbonate chemistry variable type to be used to compute CO2 in addition to pH.
Select from 'dic' 'alk' 'co3' hco3' or 'omegac'. Defaults to 'dic'.}

\item{cc2ndparm.pt}{Second carbonate chemistry prior type select from 't1' for prior defined at time step 1
with autocorrelated time series model or 'ts' for time series input by user to define prior at each time step.}

\item{cc2ndparmTS}{If 'ts' is selected for 'cc2ndparm.pt', user must provide time series R data object
consisting of 3 columns of data: age, mean value for 2nd carb chem variable, and sd for 2nd carb chem variable.}
}
\value{
Returns list 'priors_foram'. This list contains a list 'clean_pri' which contains all of the prior
information used in the model inversion, and 'psm_type' which indicates which proxy model script to use.
}
\description{
This function generates a list of objects associated with the prior distributions to be used in the mcmc
inversion. It requires users to update a sheet defining all of the prior distributions, 'priors_in_foram.R'.
It determines d11B vital effects slope and intercept for each modern species by running an mcmc inversion of each
modern species' calibration data which is routinely updated. The function also generates the temp-, sal-, and press-
dependent equilibrium constant look-up arrays which are used by the proxy system model. These arrays are updated to
reflect sensitivities to major ion concentrations of Ca, Mg and SO4 in the proxy system model. Dependencies: 
'rjags' and 'R2jags'.
}
\examples{
priors_foram(priors = "package/priors_in_foram.R", age_index = age_index, cc2ndparm.vt = 'dic',
cc2ndparm.pt = 'ts', cc2ndparmTS = cc2ndparmTS, save.output = FALSE)

}
