% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_inversion.R
\name{run_inversion}
\alias{run_inversion}
\title{Runs the MCMC inversion using groomed data}
\usage{
run_inversion(
  clean_obs = clean_obs,
  priors_foram = priors_foram,
  save.parms = c("sal", "tempC", "xca", "xmg", "xso4", "d11Bsw", "d18Osw",
    "d18Osw.local", "pco2", "dic", "pH", "press"),
  n.iter = 10000,
  n.chains = 3,
  n.burnin = 3000,
  n.thin = 1,
  parallel = FALSE
)
}
\arguments{
\item{clean_obs}{User provides the list that is returned by 'clean_foram', 'clean_phyto', 'clean_paleosol', or
'clean_plants' function. Defauts to 'clean_obs'.}

\item{save.parms}{Provide a vector of character strings indicating each model parameter for which posteriors will be
recorded in the output. Defaults to c("sal", "tempC", "xca", "xmg", "xso4", "d11Bsw", "d18Osw", "d18Osw.local",
"pco2", "dic", "pH", "press").}

\item{n.iter}{Provide the number of iterations to perform in the MCMC inversion. See 'jags' user manual for details.
Defaults to '10000'.}

\item{n.chains}{Provide the number of chains for the MCMC inversion. See 'jags' user manual for details. Defaults
to '3'.}

\item{n.burnin}{Provide the number of iterations for inversion burn-in. See 'jags' user manual for details. Defaults
to '3000'.}

\item{n.thin}{Provide the factor by which to thin the output data, i.e., '10' only saves every 10th iteration. See
'jags' user manual for details. Defaults to '1'.}

\item{parallel}{Logical. Specify TRUE if you want to enable parallel processing. Defaults to FALSE.}

\item{priors}{User provides the list that is returned by 'priors_foram', 'priors_phyto', 'priors_paleosol', or
'priors_plants' function. Defauts to 'priors_foram'.}
}
\value{
Returns list 'inv_out' with 'jout' MCMC data object, a vector of timestep ages 'ages_prox', and a list of 
the saved parameters 'save.parms'. 'jout' is the same output list that is generated when you call the 'jags' function.
}
\description{
This function loads data to send to jags, writes the proxy system model text string to be used in the inversion
and runs the jags MCMC inversion. Dependencies: 'rjags' and 'R2jags'.
}
\examples{
run_inversion(clean_obs = clean_obs, priors_foram = priors_foram,
save.parms = c("sal", "tempC", "xca", "xmg", "xso4", "d11Bsw", "d18Osw", "d18Osw.local","pco2", "dic", "pH", "press"),
n.iter = 10000, n.chains = 3, n.burnin = 3000, n.thin = 1, save.output = FALSE)

}
