% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_plot.R
\name{trace_plot}
\alias{trace_plot}
\title{Generates trace plots to examine chain convergence}
\usage{
trace_plot(inv_out = inv_out, parm = "pco2", n = 1)
}
\arguments{
\item{inv_out}{User inputs the MCMC data object that is returned by the 'run_inversion' function. Defaults to 'inv_out'.}

\item{parm}{Specify which parameter to plot. This parameter must be included in 'save.parms' argument in 'run_inversion'
function. Defaults to 'pco2'.}

\item{n}{Indicate the number of plots to step through if parameter is time-dependent. Defaults to '1'.}
}
\value{
Returns trace plot(s) to user's plot window.
}
\description{
This function generates trace plots for a indicated parameter to monitor chain convergence in the MCMC inversion.
If the parameter is time-dependent, multiple plots can be generated by increasing 'n' argument - the function will
randomly select the time step to display. Dependencies: 'mcmcplots'.
}
\examples{
trace_plot(inv_out = inv_out, parm = "pco2", n = 1)

}
