% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbchem.R
\name{carbchem}
\alias{carbchem}
\title{Marine carbonate chemistry calculator}
\usage{
carbchem(
  ccparm2 = "dic",
  ccout = "pco2",
  pH = pH,
  ccparm2.vec = ccparm2.vec,
  temp = temp,
  sal = sal,
  press = press,
  xca = xca,
  xmg = xmg,
  xso4 = xso4
)
}
\arguments{
\item{ccparm2}{User indicates which 2nd carbonate chemistry variable to use: 'dic', 'alk', 'pco2', 'co3', 'hco3', or 'omegac'.
Defaults to 'dic'.}

\item{ccout}{Specify which parameter to compute: 'dic', 'alk', 'pco2', 'co3', 'hco3', or 'omegac'. Defaults to 'co2a'.}

\item{pH}{Provide a vector of pH data in total scale. Defaults to 'pH'.}

\item{ccparm2.vec}{Provide a vector of data of the type specified by 'ccparm2'. Defaults to 'ccparm2.vec'.}

\item{temp}{Provide a vector of temperature data in degrees C. Defaults to 'temp'.}

\item{sal}{Provide a vector of salinity data. Defaults to 'sal'.}

\item{press}{Provide a vector of pressure data in bar. Defaults to 'press'.}

\item{xca}{Provide a vector of Ca ion concentration data in mmol/kg. Defaults to 'xca'.}

\item{xmg}{Provide a vector of Mg ion concentration data in mmol/kg. Defaults to 'xmg'.}

\item{xso4}{Provide a vector of SO4 ion concentration data in mmol/kg. Defaults to 'xso4'.}
}
\value{
Returns vector of values for parameter of user interest specified by 'ccout' argument.
}
\description{
This function computes any 6 of the master marine carbonate chemistry variables, plus omega for calcite, from pH data
plus an additional carbonate chem variable data set. The user must suply equal length vectors for each influencing variable
in the calculations: 'pH', 2nd carb chem variable 'ccparm2.vec', 'temp', 'sal', 'press', 'xca', 'xmg', and 'xso4'.
Units for each variable as follows: pH = total scale; ccparm2 = mmol/kg (dic, alk, co3, hco3) OR ppmv (pco2) OR unitless (omegac);
temp = deg C, sal = unitless; press = bar; elemental concentrations are in mmol/kg.
}
\examples{
carbchem(ccparm2 = "dic", ccout = "pco2", pH = pH, ccparm2.vec = ccparm2.vec, temp = temp, sal = sal, press = press,
xca = xca, xmg = xmg, xso4 = xso4)

}
