% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parms_foram_adj.R
\name{parms_foram_adj}
\alias{parms_foram_adj}
\title{Revises prior distributions and parameters}
\usage{
parms_foram_adj(parms2change = parms2change, change_values = change_values)
}
\arguments{
\item{parms2change}{Provide a vector of character strings specifying which parameters to revise. The revisable parameters are
defined in 'Details'.}

\item{change_values}{Provide a vector of values which correspond to 'priors2change' vector of parameter names.}
}
\value{
Returns list 'parms_foram_adj' which contains all of the values, revised or otherwise, for revisable parameters.
}
\description{
This function allows users to pull in the default system file which defines all user-accessible parameters, including priors,
and revise it as they see fit. This is a highly recommended step as prior distributions should reflect prior knowledge
for the specific study interval.
}
\details{
The following is a list of the revisable parameters which define the prior distributions: \cr
\cr
--- S, T, P --- \cr
\strong{'tempC.m'} and \strong{'tempC.sd'} are the mean and stdev for temperature in degrees C  \cr
\strong{'sal.m'} and \strong{'sal.sd'} are the mean and stdev for salinity  \cr
\strong{'press.m'} and \strong{'press.sd'} are the mean and stdev for presure in bar \cr
\cr
--- seawater composition --- \cr
\strong{'d11Bsw.m'} and \strong{'d11Bsw.sd'} are the mean and stdev for seawater d11B in per mille, SRM-951  \cr
\strong{'d18Osw.m'} and \strong{'d18Osw.sd'} are the mean and stdev for seawater d18O in per mille, VSMOW \cr
\strong{'xca.m'} and \strong{'xca.sd'} are the mean and stdev for Ca concentration of seawater in mmol/kg  \cr
\strong{'xmg.m'} and \strong{'xmg.sd'} are the mean and stdev for Mg concentration of seawater in mmol/kg  \cr
\strong{'xso4.m'} and \strong{'xso4.sd'} are the mean and stdev for SO4 concentration of seawater in mmol/kg  \cr
\strong{'xca.lt'} provides an option to prescribe linear change in Ca concentration of seawater as function of age, mmol/kg per kyr \cr
\strong{'xmg.lt'} provides an option to prescribe linear change in Mg concentration of seawater as function of age, mmol/kg per kyr \cr
\strong{'xso4.lt'} provides an option to prescribe linear change in SO4 concentration of seawater as function of age, mmol/kg per kyr \cr
\cr
--- diagenesis d18O correction --- \cr
\strong{'seccal'} and \strong{'seccal.sd'} are the mean and stdev for the percentage of secondary calcite \cr
\strong{'d18Oseccal'} is the estiamted d18O, per mille VPDB, of secondary calcite \cr
\cr
--- Mg/Ca temp calibration parameters --- \cr
\strong{'Hp.mean'} and \strong{'Hp.sd'} are the mean and stdev for nonlinearity of the relationship b/w shell and Mg/Casw \cr
\strong{'Bmod.mean'} and \strong{'Bmod.sd'} are the mean and stdev for modern, pre-corrected, pre-exponential constant in Mg/Ca-SST calibration \cr
\strong{'A.mean'} and \strong{'A.sd'} are the mean and stdev for the exponential constant in Mg/Ca-SST calibration \cr
\strong{'pHpccorr'} and \strong{'pHpccorrsd'} are the mean and stdev for the pH correction on Mg/Caf in percent per tenth pH unit \cr
\cr
--- d11B vital effect --- \cr
\strong{'m.custom'}, \strong{'m.customsd'}, \strong{'c.custom'}, and \strong{'c.customsd'} specify 'custom' vital effect slope and intercept  \cr
\strong{'Grub.coff'}, \strong{'Tsac.coff'}, \strong{'Ouni.coff'}, and \strong{'borate.coff'} are the 'c' intercept offsets for each modern species 'c' value; leave '0' for modern \cr
\cr
--- carboante chemistry --- \cr
\strong{'pH.u'} and \strong{'pH.l'} are the upper and lower bounds on uniform distribution for time step 1 in total scale \cr
\strong{'carbchem2.m'} and \strong{'carbchem2.sd'} are the mean and stdev for 2nd carbonate chemistry variable for time step 1*.  \cr
\cr
Carbonate chemistry variables in the following units: \cr
pH = 'total scale' equivalent  \cr
DIC = μmol/kg \cr
ALK = μmol/kg \cr
CO3 = μmol/kg \cr
HCO3 = μmol/kg \cr
\cr
*Note that variable type, 'cc2ndparm.vt', should be specified in arguments in 'foram_priors' function. These values will be used if
prior type, i.e. 'cc2ndparm.pt', is set to 't1' in 'foram_priors' argument. These values will not be used if prior type,
i.e. 'cc2ndparm.pt', is set to 'ts' in 'foram_priors' argument.
}
\examples{
parms_foram_adj(parms2change = parms2change, change_values = change_values)

}
