% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ethnorace.R
\name{predict_ethnorace}
\alias{predict_ethnorace}
\title{Predict Ethnicity/Race}
\usage{
predict_ethnorace(input_data = example_persons, bper_data = NULL, year)
}
\arguments{
\item{input_data}{The input data frame containing the individuals whose ethnorace
the user wants to predict.}

\item{bper_data}{The data list containing ethnorace conditional
probabilities. If left empty, will default to downloading directly from
Census API. Use the function \code{load_bper_data} to save this data ahead of
time.}

\item{year}{The year for which Census data will be loaded}
}
\value{
Returns the original data.frame with the additional columns for
ethnorace probabilities and predicted category.
}
\description{
Calculates posterior probabilities for individual ethnorace categories using
the Naive Bayes algorithm. Also returns highest predicted race as a new
string column in the data frame.
}
