% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_race.R
\name{predict_race}
\alias{predict_race}
\title{Predict Race}
\usage{
predict_race(df, dichotomize = FALSE)
}
\arguments{
\item{df}{A data frame containing the following column names:
\code{last_name} character, \code{first_name} character, \code{female}
numeric, \code{apartment} numeric, \code{party} character,
\code{birth_year} numeric, \code{zip} character, \code{block} character.
These columns need to all be included to run the function. If missing,
please create a column of NA's corresponding to the correct type.}

\item{dichotomize}{Indicates whether to create binary factor variables
indicating an individual's category. 'FALSE' by default.}
}
\value{
The original data.frame with the additional columns for ethnorace
probabilities and predicted category.
}
\description{
Calculates posterior probabilities for individual ethnorace categories using
the Naive Bayes algorithm. Also returns highest predicted race as a new
string column in the data frame.
}
